feat: 171 - implement live cart rendering
