+++ src/components/cart/items-wrapper.ts
@@ -1,28 +1,18 @@
-import { createA, createDiv } from '../../utils/create-elements/create-tags';
+import { createDiv } from '../../utils/create-elements/create-tags';
 import { CART } from '../../pages/cart/constants';
-import { getOrCreateCart } from '../../app/api/get-or-create-cart';
-import { createCartItem } from './item/item';
-import { createInfoMessage } from '../../shared/components/info-message';
-import { LINK_CONFIG } from '../../shared/ui-config/link';
+import { CustomEventEmitter } from '../../utils/event-emitter';
+import { renderCartItems } from '../../utils/render-cart/render-cart';
+
+export const cartEventEmitter = new CustomEventEmitter();
 
 export async function createProductsWrapper(layout: HTMLDivElement): Promise<HTMLDivElement> {
   const itemsWrapper = createDiv({ parent: layout, classes: CART.itemsWrapper });
 
-  const cart = await getOrCreateCart();
+  cartEventEmitter.subscribe('item-delete', () => {
+    renderCartItems(itemsWrapper);
+  });
 
-  if ('message' in cart) {
-    return itemsWrapper;
-  }
-
-  if (cart.lineItems.length === 0) {
-    createInfoMessage(itemsWrapper, 'Your cart is empty. Start shopping and fill it with some amazing items!');
-    const catalogLink = createA(LINK_CONFIG.catalog);
-    itemsWrapper.append(catalogLink);
-  } else {
-    for (const item of cart.lineItems) {
-      createCartItem(item, itemsWrapper);
-    }
-  }
+  renderCartItems(itemsWrapper);
 
   return itemsWrapper;
 }
