+++ src/utils/render-cart/render-cart.ts
@@ -1,0 +1,25 @@
+import { getOrCreateCart } from '../../app/api/get-or-create-cart';
+import { createInfoMessage } from '../../shared/components/info-message';
+import { createA } from '../create-elements/create-tags';
+import { LINK_CONFIG } from '../../shared/ui-config/link';
+import { createCartItem } from '../../components/cart/item/item';
+
+export async function renderCartItems(itemsWrapper: HTMLDivElement): Promise<void> {
+  itemsWrapper.replaceChildren();
+
+  const cart = await getOrCreateCart();
+
+  if ('message' in cart) {
+    return;
+  }
+
+  if (cart.lineItems.length === 0) {
+    createInfoMessage(itemsWrapper, 'Your cart is empty. Start shopping and fill it with some amazing items!');
+    const catalogLink = createA(LINK_CONFIG.catalog);
+    itemsWrapper.append(catalogLink);
+  } else {
+    for (const item of cart.lineItems) {
+      createCartItem(item, itemsWrapper);
+    }
+  }
+}
